Return the number of participants when getting returning trips

+++ src/main/java/com/nik/tripfinder/repositories/ReservationRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
-    long countByTripId(Long tripId);
+    Integer countByTripId(Long tripId);
     Optional<Reservation> findReservationByCustomerCustomerIdAndTripId(Integer customerId, Long tripId);
     List<Reservation> findReservationsByTripId(Long tripId);
     List<Reservation> findReservationsByCustomerCustomerId(Integer customerId);
